fix image in venue/menu/list
