+++ src/components/Venue/Menu/index.js
@@ -32,6 +32,7 @@ const MenuList = () => {
         setLoading(true)
         const res = await productsVenueList(auth.user.signInUserSession.idToken.jwtToken)
         if(isActive){
+          console.log(res.data)
           setItems(res.data)
           setLoading(false)
         }
